refactor: 도메인간 서비스 의존관계 UseCase로 추상화
