+++ src/main/kotlin/com/server/dpmcore/cohort/application/CohortQueryService.kt
@@ -3,7 +3,8 @@ package com.server.dpmcore.cohort.application
 import com.server.dpmcore.cohort.application.config.CohortProperties
 import com.server.dpmcore.cohort.domain.exception.CohortNotFoundException
 import com.server.dpmcore.cohort.domain.model.CohortId
-import com.server.dpmcore.cohort.domain.port.CohortPersistencePort
+import com.server.dpmcore.cohort.domain.port.inbound.CohortQueryUseCase
+import com.server.dpmcore.cohort.domain.port.outbount.CohortPersistencePort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -12,8 +13,8 @@ import org.springframework.transaction.annotation.Transactional
 class CohortQueryService(
     private val cohortPersistencePort: CohortPersistencePort,
     private val cohortProperties: CohortProperties,
-) {
-    fun getLatestCohortId(): CohortId =
+) : CohortQueryUseCase {
+    override fun getLatestCohortId(): CohortId =
         cohortPersistencePort.findCohortIdByValue(cohortProperties.value)
             ?: throw CohortNotFoundException()
 }
