+++ src/main/kotlin/com/server/dpmcore/cohort/domain/port/inbound/CohortQueryUseCase.kt
@@ -1,0 +1,7 @@
+package com.server.dpmcore.cohort.domain.port.inbound
+
+import com.server.dpmcore.cohort.domain.model.CohortId
+
+interface CohortQueryUseCase {
+    fun getLatestCohortId(): CohortId
+}
